# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CNC Research Platform Product API
  version: 1.0.0
  description: |
    Authenticated Product API for creating CNC research experiments, queueing
    episodes, polling their lifecycle, and reading persisted contract-v2 events.

    Starting or retrying an episode only queues database work. Product API
    request handlers never execute the research engine.

    Clients may send `X-Request-Id` using 1-128 letters, numbers, dots,
    underscores, colons, or hyphens. The API echoes it in the response; when
    omitted or invalid, the API generates a UUID request id.
servers:
  - url: http://localhost:5000/api/v1
    description: Local backend
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /decision-cases:
    post:
      summary: Create a case without executing a planner
      description: >-
        OPERATOR, ENGINEER or ADMIN. Key scoped to actor and factory. Replays return the original 202 response. Server
        owns IDs, actor, timestamp, revision and status.
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '^[!-~]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCaseCreateRequest'
      responses:
        '202':
          description: 'Persisted, sanitized response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionCaseResponse'
          headers:
            Location:
              schema:
                type: string
            Idempotency-Replayed:
              schema:
                const: 'true'
        '400':
          description: 'Invalid strict body, trigger, idempotency key or cursor'
        '401':
          description: Authentication required
        '403':
          description: Role cannot create cases
        '404':
          description: Unknown or inaccessible resource
        '409':
          description: 'IDEMPOTENCY_KEY_REUSED, SNAPSHOT_CONFLICT or PLAN_VERSION_CONFLICT'
        '500':
          description: Operation failed atomically; no private details exposed
  '/decision-cases/{id}':
    get:
      summary: Read generated status and consistent current head
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Persisted, sanitized response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionCaseResponse'
        '401':
          description: Authentication required
        '404':
          description: Unknown or inaccessible resource
  '/decision-cases/{id}/events':
    get:
      summary: Read sanitized events ascending by sequence; limit + 1 determines has_more
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: after_sequence
          schema:
            type: integer
            minimum: 0
            maximum: 2147483647
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: 'Persisted, sanitized response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionCaseEventsResponse'
        '400':
          description: 'Invalid strict body, trigger, idempotency key or cursor'
        '401':
          description: Authentication required
        '404':
          description: Unknown or inaccessible resource
...
  /maintenance/decisions:
    get:
      tags: [Maintenance]
      operationId: listMaintenanceDecisionCandidates
      summary: List machine-level maintenance candidates from persisted simulator events
      description: >-
        Each complete observation, policy recommendation and step result produces one candidate per machine
        for the episode's current attempt. ADMIN can read all experiments; other users can read their own.
        Results are ordered by severity (CRITICAL first), estimated cost of one-step delay (highest first),
        creation time (newest first), then candidate id. The review status belongs to the shared joint
        recommendation; decisionId is null until POST /maintenance/decisions opens it for human review.
        Simulator results are research outcomes, not maintenance execution acknowledgements.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_REVIEW, APPROVED, OVERRIDDEN, REJECTED]
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW]
        - name: experiment
          in: query
          description: Experiment UUID.
          schema:
            type: string
            format: uuid
        - name: machine
          in: query
          description: Exact machine identifier.
          schema:
            $ref: '#/components/schemas/Identifier'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Filtered and prioritized candidates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceCandidateListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    post:
      tags:
        - Maintenance
      operationId: createMaintenanceDecision
      summary: Open a recommendation for human review
      description: >-
        Requires authentication. ADMIN can access all decisions; other roles access decisions belonging to their
        own experiments. VIEWER can only read. No operation executes machine actions. OPERATOR, ENGINEER, ADMIN
        only. Idempotently returns the existing decision for this recommendation without resetting its status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - recommendationId
              properties:
                recommendationId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Pending or existing decision
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/MaintenanceDecision'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  '/maintenance/decisions/{id}':
    get:
      tags: [Maintenance]
      operationId: getMaintenanceDecisionCandidate
      summary: Get one machine-level maintenance candidate
      description: >-
        id is recommendation UUID, a tilde, and the machine identifier. The candidate is derived from the
        current episode attempt's persisted observation, recommendation and step result. Owner scope is
        identical to the list endpoint. The id differs from the UUID used by review/history endpoints.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F-]{36}~[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
      responses:
        '200':
          description: Candidate and its source snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceCandidateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pdm_token
      description: HttpOnly browser session cookie issued by POST /auth/login.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing, malformed, expired, or inactive-user authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: The authenticated role does not permit this mutation. Ownership mismatches are intentionally returned as 404.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have permission to access this resource
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Request body, path, or query validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: horizon_steps must be greater than or equal to 1
              details:
                - path: environmentConfig.horizon_steps
                  code: too_small
                  message: Must be greater than or equal to 1
    Conflict:
      description: A uniqueness or resource-state conflict prevented the operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: CONFLICT
              message: The requested operation conflicts with current state
...
    DatabaseUnavailable:
      description: PostgreSQL is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: DB_UNAVAILABLE
              message: Database unavailable. Check DATABASE_URL and make sure Postgres is running.
  schemas:
    CaseContractDecisionCaseStatusResponse:
      additionalProperties: false
      properties:
        committed_schedule_id:
          anyOf:
            - maxLength: 128
              minLength: 1
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
              type: string
            - type: 'null'
          default: null
          title: Committed Schedule Id
        created_at:
          format: date-time
          title: Created At
          type: string
        decision_case_id:
          maxLength: 128
          minLength: 1
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
          title: Decision Case Id
          type: string
        error_code:
          anyOf:
            - maxLength: 128
              minLength: 1
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
              type: string
            - type: 'null'
          default: null
          title: Error Code
        mode:
          $ref: '#/components/schemas/CaseContractDecisionMode'
        recommendation_id:
          anyOf:
            - maxLength: 128
              minLength: 1
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
              type: string
            - type: 'null'
          default: null
          title: Recommendation Id
        schema_version:
          const: '3.0'
          default: '3.0'
          title: Schema Version
          type: string
        snapshot_id:
          maxLength: 128
          minLength: 1
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
          title: Snapshot Id
          type: string
        status:
          $ref: '#/components/schemas/CaseContractDecisionCaseStatus'
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
        - decision_case_id
        - mode
        - status
        - snapshot_id
        - created_at
        - updated_at
      title: DecisionCaseStatusResponse
      type: object
    CaseContractDecisionMode:
      enum:
        - LIVE
        - SIMULATION_ONLY
      title: DecisionMode
      type: string
    CaseContractDecisionCaseStatus:
      enum:
        - CREATED
        - ANALYZING
        - GENERATING
        - VALIDATING
        - EXPLAINING
        - AWAITING_APPROVAL
        - APPROVED
        - MODIFIED
        - REJECTED
        - COMMITTED
        - FAILED
        - CANCELLED
      title: DecisionCaseStatus
      type: string
    CaseContractTimeWindow:
      additionalProperties: false
      properties:
        end_at:
          format: date-time
          title: End At
          type: string
        start_at:
          format: date-time
          title: Start At
          type: string
      required:
        - start_at
        - end_at
      title: TimeWindow
      type: object
    CaseContractPlanningConfig:
      additionalProperties: false
      properties:
        allowed_strategy_ids:
          items:
            maxLength: 128
            minLength: 1
            pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
            type: string
          title: Allowed Strategy Ids
          type: array
        base_seed:
          default: 0
          minimum: 0
          title: Base Seed
          type: integer
        candidate_limit:
          default: 3
          maximum: 5
          minimum: 1
          title: Candidate Limit
          type: integer
        horizon_minutes:
          minimum: 1
          title: Horizon Minutes
          type: integer
        objective_weights:
          $ref: '#/components/schemas/CaseContractObjectiveWeights'
        simulation_runs:
          default: 100
          maximum: 10000
          minimum: 1
          title: Simulation Runs
          type: integer
        solver_timeout_seconds:
          default: 30
          maximum: 300
          minimum: 1
          title: Solver Timeout Seconds
          type: integer
      required:
        - horizon_minutes
      title: PlanningConfig
      type: object
    CaseContractObjectiveWeights:
      additionalProperties: false
      properties:
        maintenance:
          default: 1
          minimum: 0
          title: Maintenance
          type: number
        production:
          default: 1
          minimum: 0
          title: Production
          type: number
        reliability:
          default: 1
          minimum: 0
          title: Reliability
          type: number
        schedule_stability:
          default: 0.25
          minimum: 0
          title: Schedule Stability
          type: number
        workforce:
          default: 1
          minimum: 0
          title: Workforce
          type: number
      title: ObjectiveWeights
      type: object
    DecisionCaseCreateRequest:
      type: object
      additionalProperties: false
      required:
        - factory_id
        - schema_version
        - expected_snapshot_id
        - expected_plan_version
        - request
      properties:
        factory_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
        schema_version:
          const: '3.0'
        expected_snapshot_id:
          type: string
          minLength: 1
          maxLength: 128
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
        expected_plan_version:
          type: integer
          minimum: 0
          maximum: 2147483647
        request:
          type: object
          additionalProperties: false
          required:
            - mode
            - trigger
            - planning_config
          properties:
            mode:
              $ref: '#/components/schemas/CaseContractDecisionMode'
            trigger:
              oneOf:
                - additionalProperties: false
                  properties:
                    alert_threshold:
                      maximum: 1
                      minimum: 0
                      title: Alert Threshold
                      type: number
                    failure_probability:
                      maximum: 1
                      minimum: 0
                      title: Failure Probability
                      type: number
                    machine_id:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Machine Id
                      type: string
                    type:
                      const: HEALTH_ALERT
                      default: HEALTH_ALERT
                      title: Type
                      type: string
                  required:
                    - machine_id
                    - failure_probability
                    - alert_threshold
                  title: HealthAlertTrigger
                  type: object
                - additionalProperties: false
                  properties:
                    failure_code:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Failure Code
                      type: string
                    machine_id:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Machine Id
                      type: string
                    type:
                      const: MACHINE_FAILURE
                      default: MACHINE_FAILURE
                      title: Type
                      type: string
                  required:
                    - machine_id
                    - failure_code
                  title: MachineFailureTrigger
                  type: object
                - additionalProperties: false
                  properties:
                    reason:
                      maxLength: 256
                      minLength: 1
                      title: Reason
                      type: string
                    technician_id:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Technician Id
                      type: string
                    type:
                      const: TECHNICIAN_UNAVAILABLE
                      default: TECHNICIAN_UNAVAILABLE
                      title: Type
                      type: string
                    unavailable_window:
                      $ref: '#/components/schemas/CaseContractTimeWindow'
                  required:
                    - technician_id
                    - unavailable_window
                    - reason
                  title: TechnicianUnavailableTrigger
                  type: object
                - additionalProperties: false
                  properties:
                    maintenance_request_id:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Maintenance Request Id
                      type: string
                    type:
                      const: MAINTENANCE_REQUESTED
                      default: MAINTENANCE_REQUESTED
                      title: Type
                      type: string
                  required:
                    - maintenance_request_id
                  title: MaintenanceRequestedTrigger
                  type: object
                - additionalProperties: false
                  properties:
                    reason:
                      maxLength: 4000
                      minLength: 1
                      title: Reason
                      type: string
                    type:
                      const: MANUAL_REPLAN
                      default: MANUAL_REPLAN
                      title: Type
                      type: string
                  required:
                    - reason
                  title: ManualReplanTrigger
                  type: object
                - additionalProperties: false
                  properties:
                    assumptions:
                      additionalProperties: true
                      title: Assumptions
                      type: object
                    scenario_id:
                      maxLength: 128
                      minLength: 1
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                      title: Scenario Id
                      type: string
                    type:
                      const: WHAT_IF
                      default: WHAT_IF
                      title: Type
                      type: string
                  required:
                    - scenario_id
                    - assumptions
                  title: WhatIfTrigger
                  type: object
            planning_config:
              $ref: '#/components/schemas/CaseContractPlanningConfig'
    DecisionCaseResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          const: true
        data:
          $ref: '#/components/schemas/CaseContractDecisionCaseStatusResponse'
        meta:
          type: object
          required:
            - factory_id
            - base_plan_version
            - case_revision
            - current_context
            - stale
          properties:
            factory_id:
              type: string
              minLength: 1
              maxLength: 128
              pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
            base_plan_version:
              type: integer
              minimum: 0
            case_revision:
              const: 1
            stale:
              type: boolean
            current_context:
              anyOf:
                - type: object
                  required:
                    - snapshot_id
                    - plan_version
                  properties:
                    snapshot_id:
                      type: string
                      minLength: 1
                      maxLength: 128
                      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
                    plan_version:
                      type: integer
                      minimum: 0
                - type: 'null'
    DecisionCaseHistoryEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - case_id
        - sequence
        - type
        - actor
        - occurred_at
        - payload
      properties:
        id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 1
        type:
          type: string
        actor:
          type: object
          required:
            - kind
            - id
          properties:
            kind:
              const: HUMAN
            id:
              type: string
              format: uuid
        occurred_at:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: false
          properties:
            status:
              const: CREATED
            revision:
              const: 1
    DecisionCaseEventsResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/DecisionCaseHistoryEvent'
        meta:
          type: object
          required:
            - has_more
            - next_after_sequence
          properties:
            has_more:
              type: boolean
            next_after_sequence:
              type: integer
              minimum: 0
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [path, code, message]
                properties:
                  path:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9._:-]+$"
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
...
    MaintenanceCandidate:
      type: object
      additionalProperties: false
      required: [id, decisionId, status, machineId, toolId, jobId, source, recommendedAction, severity, priority, risk, resources, predictedCost, costOfDelay, rationale, rationaleSource, result, createdAt]
      properties:
        id:
          type: string
          description: Stable recommendation UUID followed by ~ and the machine identifier.
        decisionId:
          type: [string, 'null']
          format: uuid
          description: Shared human review decision UUID, if opened.
        status:
          type: string
          enum: [PENDING_REVIEW, APPROVED, OVERRIDDEN, REJECTED]
          description: Status of the joint recommendation's human review; PENDING_REVIEW before a review row exists.
        machineId:
          $ref: '#/components/schemas/Identifier'
        toolId:
          $ref: '#/components/schemas/Identifier'
        jobId:
          type: [string, 'null']
        source:
          type: object
          additionalProperties: false
          required: [experimentId, episodeId, episodeIndex, attempt, step, observationId, recommendationId, policyId, resultId]
          properties:
            experimentId: { type: string, format: uuid }
            episodeId: { type: string, format: uuid }
            episodeIndex: { type: integer, minimum: 0 }
            attempt: { type: integer, minimum: 1 }
            step: { type: integer, minimum: 0 }
            observationId: { type: string, format: uuid }
            recommendationId: { type: string, format: uuid }
            policyId: { type: string, format: uuid }
            resultId: { type: string, format: uuid }
        recommendedAction:
          type: string
          enum: [REPLACE_NOW, SCHEDULE_REPLACEMENT, DEFER]
          description: REPLACE with spares maps to REPLACE_NOW, REPLACE without spares to SCHEDULE_REPLACEMENT, CONTINUE to DEFER.
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
          description: >-
            CRITICAL at one-step failure probability >= 0.5 or wear >= threshold; HIGH at probability >= 0.25,
            wear >= 80% of threshold or policy REPLACE; MEDIUM at probability >= 0.1 or wear >= 50%; else LOW.
        priority:
          type: integer
          minimum: 1
          maximum: 4
          description: 1=CRITICAL, 2=HIGH, 3=MEDIUM, 4=LOW.
        risk:
          type: object
          additionalProperties: false
          required: [predictedRulSteps, failureProbabilityNextStep, observedWearUm, posteriorMedianWearUm, failureThresholdUm, condition]
          properties:
            predictedRulSteps:
              type: [integer, 'null']
              minimum: 0
              description: Median of the persisted RUL distribution; null if survival beyond horizon exceeds 50%.
            failureProbabilityNextStep:
              type: number
              minimum: 0
              maximum: 1
              description: Probability mass at RUL support steps 0 and 1.
            observedWearUm: { type: number, minimum: 0 }
            posteriorMedianWearUm: { type: number, minimum: 0 }
            failureThresholdUm: { type: number, exclusiveMinimum: 0 }
            condition:
              type: object
              additionalProperties: false
              required: [conditionId, loadClass]
              properties:
                conditionId: { $ref: '#/components/schemas/Identifier' }
                loadClass: { type: string, enum: [LIGHT, NOMINAL, HEAVY] }
        resources:
          type: object
          additionalProperties: false
          required: [sparesAvailable, inventoryCapacity]
          properties:
            sparesAvailable: { type: integer, minimum: 0 }
            inventoryCapacity: { type: integer, minimum: 0 }
        predictedCost:
          type: object
          additionalProperties: false
          required: [expected, cvar, scope]
          properties:
            expected: { type: [number, 'null'], minimum: 0 }
            cvar: { type: [number, 'null'], minimum: 0 }
            scope: { const: JOINT_RECOMMENDATION }
          description: Policy estimates for the whole joint action, not a machine-specific allocation.
        costOfDelay:
          type: object
          additionalProperties: false
          required: [amount, basis]
          properties:
            amount:
              type: [number, 'null']
              minimum: 0
              description: One-step failure probability multiplied by configured failure cost; null if cost is absent.
            basis: { const: ONE_STEP_FAILURE_EXPOSURE }
          description: Derived exposure proxy, not a counterfactual policy forecast.
        rationale:
          type: string
        rationaleSource:
          const: DERIVED_FROM_PERSISTED_EVENTS
          description: Policy payload has no free-text rationale; this text is composed from action and risk/resource snapshot.
        result:
          type: object
          additionalProperties: false
          required: [outcome, incurredCost]
          properties:
            outcome: { type: string, enum: [CONTINUED, REPLACED, WAITING_FOR_SPARE, FAILED] }
            incurredCost: { type: number, minimum: 0 }
          description: Persisted simulator outcome, not a maintenance execution acknowledgement.
        createdAt:
          type: string
          format: date-time
    MaintenanceCandidateResponse:
      type: object
      additionalProperties: false
      required: [success, data]
      properties:
        success: { const: true }
        data: { $ref: '#/components/schemas/MaintenanceCandidate' }
    MaintenanceCandidateListResponse:
      type: object
      additionalProperties: false
      required: [success, data, meta]
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/MaintenanceCandidate' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    MaintenanceDecision:
      type: object
      required:
        - id
        - recommendationId
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        recommendationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING_REVIEW
            - APPROVED
            - OVERRIDDEN
            - REJECTED
        createdAt:
          type: string
          format: date-time
...